# public/openapi.yaml
openapi: 3.0.3
info:
  title: Pixel Night API
  description: |
    API interne de Pixel Night — application de soirées cinéma collaboratives.

    Les endpoints de vote sont publics (pas d'authentification) ; les endpoints d'administration
    requièrent une session Supabase valide (cookie `sb-*` posé par l'auth Next.js).

    **Identifiant votant anonyme** : `voter_id` est un UUID généré côté client et stocké dans
    le `localStorage` du navigateur. Il permet de garantir l'unicité du vote sans compte.
  version: 1.3.0
  contact:
    url: https://github.com/Jagostini/pixel-night
  license:
    name: MIT + Commons Clause
    url: https://github.com/Jagostini/pixel-night/blob/main/LICENSE
servers:
  - url: /api
    description: Serveur courant
...
paths:
...
  /soirees/{id}/finalize-film:
    post:
      tags: [Administration]
      summary: Finaliser le vote de film
      description: |
        Désigne le film gagnant (en cas d'égalité, tirage au sort).
        Passe la soirée en phase `completed`.
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/SoireeId'
      responses:
        '200':
          description: Film gagnant désigné
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  winning_film_tmdb_id:
                    type: integer
                  winning_film_title:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /soirees/{id}/fetch-films-discover:
    post:
      tags: [Administration]
      summary: Récupérer les films via TMDb Discover (genres)
      description: |
        Interroge TMDb Discover avec les `genre_ids` du thème gagnant.
        Récupère 5 pages aléatoires parmi les 10 premières pour de la variété.

        Si le thème n'a pas de `genre_ids` configurés, effectue un repli automatique
        sur la recherche par mots-clés (identique à `fetch-films`).

        **Filtres Discover** :
        - `vote_count.gte=100`
        - `vote_average.gte=6`
        - `sort_by=popularity.desc`
      security:
        - SupabaseSession: []
      parameters:
        - $ref: '#/components/parameters/SoireeId'
      responses:
        '200':
          description: Films insérés
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  used_discover:
                    type: boolean
                    description: True si TMDb Discover a été utilisé, false si repli mots-clés
        '400':
          description: Pas de thème gagnant défini
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Aucun film trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Cookie de session Supabase posé automatiquement après connexion
  parameters:
    SoireeId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID de la soirée
  responses:
    BadRequest:
      description: Paramètres manquants ou incorrects
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Session non authentifiée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Non autorise
    Forbidden:
      description: Accès refusé — l'utilisateur n'est pas le créateur de la soirée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Non autorise
    InternalError:
      description: Erreur serveur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...